+++ common/image_preprocessor.py
