+++ singleshot/dipole_trap_single_shot.py
@@ -1,9 +1,10 @@
-from common.analysis_config import kinetix_system, BulkGasAnalysisConfig
-from common.bulk_gas_analysis import BulkGasPreprocessor
-from common.bulk_gas_statistics import BulkGasStatistician
-from common.image import ROI
-from common.plot_config import PlotConfig
-from matplotlib import pyplot as plt
+import matplotlib.pyplot as plt
+
+from analysislib.common.analysis_config import kinetix_system, BulkGasAnalysisConfig
+from analysislib.common.bulk_gas_analysis import BulkGasPreprocessor
+from analysislib.common.bulk_gas_statistics import BulkGasStatistician
+from analysislib.common.image import ROI
+from analysislib.common.plot_config import PlotConfig
 
 
 analysis_config = BulkGasAnalysisConfig(
